few more tweeks
