--- Autodan/src/pages/MerchTool/MerchandisePages/MtCommonToMerchandiesDetailsPage.cs
@@ -1,12 +1,0 @@
-using OpenQA.Selenium;
-
-namespace Autodan.pages.MerchTool.MerchandisePages
-{
-    public interface IMtCommonToMerchandiseDetailsPage
-    {
-        void VerifyElements();
-        void VerifyPresentation();
-        IWebElement BreadCrumbDetails { get; set; }
-        IWebElement Legend { get; set; }
-    }
-}
